Implementada interfaz de rating sin funcionalidad

+++ database/migrations/2020_02_24_111710_create_students_table.php
@@ -26,7 +26,7 @@ class CreateStudentsTable extends Migration
             $table->string('password');
             $table->string('studies');
             $table->integer('course');
-            $table->integer('puntuation');
+            $table->double('puntuation');
         });
     }
